report-management - move _queries into bark/queries
