+++ src/lib/bark/queries/select-pending-appointment-exists.ts
